refactor: separate the handles
