+++ https/account/request.go
@@ -1,0 +1,8 @@
+package account
+
+type CreateRequest struct {
+	Name    string `json:"name"`
+	CPF     string `json:"cpf"`
+	Secret  string `json:"secret"`
+	Balance int    `json:"balance"`
+}
