+++ https/authentication/request.go
@@ -1,0 +1,6 @@
+package authentication
+
+type LoginRequest struct {
+	CPF    string `json:"cpf"`
+	Secret string `json:"secret"`
+}
