+++ https/authentication/response.go
@@ -1,0 +1,5 @@
+package authentication
+
+type TokenResponse struct {
+	Token string `json:"token"`
+}
