+++ https/error.go
@@ -1,0 +1,5 @@
+package https
+
+type ErrorsResponse struct {
+	Errors string `json:"errors"`
+}
